fix(web-demo): 清偿全量 eslint 债务并根治测试基建两处病灶
- eslint --fix 批量收敛 import/order 等 74 处可自动修复项
- 手工修复 18 处 error: 空 no-op 回调改 () => undefined、PdfViewer
  悬浮 Promise 加 void、xlsx 引用去 any 改真实类型、formatArgs 不可
  序列化分支、isTesting 空值合并、setInterval 异步回调 void 包装、
  desktop 后端启动失败 catch 落 debug 日志说明吞咽原因
- thread-stream-cache.test: useSubtaskContext mock 每渲染新建 setTasks
  身份, 与 useThreadStream 重置 effect 的依赖链叠加成无界重渲染循环,
  worker 堆内存耗尽崩溃; 提升 mock 至 vi.hoisted 稳定引用后 8/8 通过
- token-usage-budget-settings.test: readQilin 根目录误指向 Python 包
  qilin/ 导致 ENOENT, 改读仓库根 app/gateway/routers/config_router.py
- 验证: pnpm test 314/314 全绿, pnpm run check (eslint+typecheck) 通过

+++ web-demo/src/components/better-sidebar/viewers/ImageViewer.tsx
@@ -1,13 +1,14 @@
 "use client";
 import { useEffect, useMemo } from "react";
 import type { FC } from "react";
+
 import type { FileViewerProps } from "@/core/sidebar/protocol";
 
 const ImageViewer: FC<FileViewerProps> = ({ entry, content }) => {
   const url = useMemo(() => {
     if (typeof content === "string") return null; // raw URL not used
     // Copy into an ArrayBuffer-backed view (TS 5.9: Uint8Array<ArrayBufferLike> is not a BlobPart).
-    const blob = new Blob([new Uint8Array(content as Uint8Array)], { type: entry.mime ?? "application/octet-stream" });
+    const blob = new Blob([new Uint8Array(content)], { type: entry.mime ?? "application/octet-stream" });
     return URL.createObjectURL(blob);
   }, [content, entry.mime]);
   useEffect(() => () => { if (url) URL.revokeObjectURL(url); }, [url]);
